pyproject.toml
[project]
name = "face-detection"
version = "0.1.0"
description = "Mobile face detector trained on CelebA dataset"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "pandas>=3.0.1",
    "gradio>=6.9.0",
    "opencv-python>=4.13.0.92",
    "litert-torch>=0.8.0",
    "tensorflow>=2.21.0",
    "torchao==0.8.0",
    "onnx>=1.20.1",
    "onnxruntime>=1.24.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"

[dependency-groups]
dev = [
    "torchvision>=0.21.0",
]
